Public model sever validation

+++ routes/controllers/QuestionController.js
@@ -1,18 +1,15 @@
 const {Question,ValidateQuestion} = require("../../models/Question");
+const {Attempt} = require("../../models/Attempt");
 const {paginateModel,paginateArray} = require("../../utility/Pagination");
 
 
-async function createQuestion(req,res) {    
-    const {error}=ValidateQuestion(req.body);         
-    if(error)return res.status(400).send(error.details[0].message);
-   
-    
+async function createQuestion(req,res) {        
     try{
     const question = new Question(req.body);
-    const isSet = setCorrectOptionIndex(req.body)
-    if(isSet != -1){
-        return -1 
-    }
+    const isSet = setCorrectOptionIndex(question)
+    if(isSet != null){
+        return isSet 
+    }    
     await question.save();
     return question
       }catch(err){
@@ -23,14 +20,32 @@ async function createQuestion(req,res) {
 
 
 
+
+
+  function setIndex(question){    
+    question.Options.forEach((element,index) => {
+      question.Options[index].Index  = index  
+      console.log(question.Options[index])
+    });
+  }
+
+
   function setCorrectOptionIndex(question){
+    setIndex(question)
     const correctOption = question.Options.filter(function(item){
+      console.log(1)
         return item.IsCorrect == true
         })
-        if (correctOption.length  != 0){
-        question.CorrectOptionIndex = correctOption[0].index
-        } else if (correctOption.length > 1){
-         return -1
+
+        if (correctOption.length > 1){
+         return {message:"Multiple Correct Options",code:-1}
+        }else  if (correctOption.length  == 0){
+          return {message:"No Correct Options",code:-1}
+        }else if (correctOption.length == 1){
+          console.log(correctOption[0].Index)
+          question.CorrectOptionIndex = correctOption[0].Index
+        }else{
+          return {message:"Unexpected Error",code:-1}
         }
   }
 
@@ -62,7 +77,6 @@ async function createQuestion(req,res) {
         }        
         
       const question = await Question.findOneAndUpdate({"_id":req.params.questionId},req.body,{new:true});
-      console.log(question);
       return question;
     }catch{
       //logger
@@ -72,8 +86,16 @@ async function createQuestion(req,res) {
   
     async function deleteQuestion(req,res) {
         try{
-        const question = await Question.findByIdAndDelete(req.params.questionId);
-        return question;
+          const attempts = Attempt.find({"QuestionsAttempted.question": req.params.questionId})
+          console.log(attempts)
+        if (attempts == null){         
+          const question = await Question.findByIdAndDelete(req.params.questionId);
+          return question;
+        }else{
+          return {message:"This Question has already been attempted and cannot be deleted,Delete the Test(s) and Attempt(s) associated with this Question",code: -1}
+        }
+
+        
         }catch{
         //logger
         }
